Assert on connectivity success
