+++ test/Altinn.App.Integration.Tests/_testapps/_shared/ConnectivityDiagnostics.cs
@@ -123,13 +123,3 @@ public static class ConnectivityDiagnostics
         return app;
     }
 }
-
-public class ConnectivityResult
-{
-    public bool Success { get; set; }
-    public int StatusCode { get; set; }
-    public string Url { get; set; } = string.Empty;
-    public string? ResponseContent { get; set; }
-    public string Message { get; set; } = string.Empty;
-    public string? Exception { get; set; }
-}
