+++ test/Altinn.App.Integration.Tests/_testapps/_shared/ConnectivityResult.cs
@@ -1,0 +1,15 @@
+using System;
+
+#nullable enable
+
+namespace TestApp.Shared;
+
+public class ConnectivityResult
+{
+    public bool Success { get; set; }
+    public int StatusCode { get; set; }
+    public string Url { get; set; } = string.Empty;
+    public string? ResponseContent { get; set; }
+    public string Message { get; set; } = string.Empty;
+    public string? Exception { get; set; }
+}
